+++ doc/changes.rst
@@ -11,12 +11,14 @@ fiberassign change log
 * Update ReadTheDocs configuration (PR `#459`_).
 * Add Assignment.check_avail_collisions(tile, all_matches=False) to check potential assignments for collisions (PR `#454`_).
 * uses the focalplane radius from DESIMODEL instead of the hardcoded version in fiberassign (PR `#453`_).
+* speeds up the Hardware constructor's search for neighboring positioners by pre-sorting in one dimension first (PR `#460`_).
 
 .. _`#453`: https://github.com/desihub/fiberassign/pull/453
 .. _`#454`: https://github.com/desihub/fiberassign/pull/454
 .. _`#459`: https://github.com/desihub/fiberassign/pull/459
 .. _`#457`: https://github.com/desihub/fiberassign/pull/457
 .. _`#458`: https://github.com/desihub/fiberassign/pull/458
+.. _`#460`: https://github.com/desihub/fiberassign/pull/460
 
 5.7.1 (2023-02-06)
 ------------------